+++ src/features/XIT/ACT/CreateActionPackage.vue
@@ -4,6 +4,7 @@ import SectionHeader from '@src/components/SectionHeader.vue';
 import Active from '@src/components/forms/Active.vue';
 import TextInput from '@src/components/forms/TextInput.vue';
 import Commands from '@src/components/forms/Commands.vue';
+import { isValidPackageName } from '@src/features/XIT/ACT/utils';
 
 const { onCreate } = defineProps<{ onCreate: (name: string) => void }>();
 
@@ -11,19 +12,15 @@ const emit = defineEmits<{ (e: 'close'): void }>();
 
 const name = ref('');
 const nameError = ref(false);
-watch(name, () => (nameError.value = !hasValidChars()));
+watch(name, () => (nameError.value = !isValidPackageName(name.value)));
 
 function onCreateClick() {
-  if (name.value.length === 0 || !hasValidChars()) {
+  if (name.value.length === 0 || !isValidPackageName(name.value)) {
     nameError.value = true;
     return;
   }
   onCreate(name.value);
   emit('close');
-}
-
-function hasValidChars() {
-  return /^[0-9a-zA-Z.-]*$/.test(name.value);
 }
 </script>